skillNexus/requirements.txt
django==4.2.0
djangorestframework
pyyaml
requests
django-cors-headers
djangorestframework-simplejwt
Pillow
django-rest-swagger
drf-yasg
drf-spectacular
